Add utility getV3ImportEqualsCode
